Fixed keyboard lockup during PC cold start (when power applied)
based on @Hamberthm fixes for esp-32. See issue #21

Hamberthm's comments:
* On my 486 PC, ACKs were being missed by the host controller during LED
set command handling. This made the process to end and the controller
to ignore the keyboard after that command. This makes it stable.

* Testing in my Toshiba 205CDS with a combined port, DEVICE ID command was
failing to write the ID properly. Added some loops to shield time to the
controller and get them properly.

* The interface was missing some commands in one of my sytems because of
waiting to send the 0xAA "BAT test success". BAT success should always
be sent a couple of houndred of milliseconds after power up,
regarding state.